all: fix cov reports about must-be-positive arguments
2. negative_return_fn: Function open(new_file, 577, 438) returns a negative number.
3. assign: Assigning: fd = open(new_file, 577, 438).
CID 1440489 (#1 of 1): Argument cannot be negative (NEGATIVE_RETURNS)
4. negative_returns: fd is passed to a parameter that cannot be negative.

CID 1441085, CID 1441390, CID 1441425, CID 1441479, CID 1441522,
CID 1441632, CID 1441692, CID 1441765

+++ exch/exchange_emsmdb/ftstream_parser.cpp
@@ -859,7 +859,7 @@ BOOL ftstream_parser_write_buffer(
 static BOOL ftstream_parser_truncate_fd(
 	FTSTREAM_PARSER *pstream)
 {
-	uint32_t len;
+	ssize_t len;
 	char buff[0x10000];
 	
 	if (0 == pstream->offset) {
